Safari-611.1.10/LayoutTests/svg/animations/animate-color-calcMode-discrete-expected.txt

SVG 1.1 dynamic animation tests

Test calcMode discrete with from-to animation on colors. You should see a green 100x100 rect and only PASS messages

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS colorComponents[1] is 128
PASS colorComponents[2] is 255
PASS colorComponents[3] is 255
PASS colorComponents[1] is 255
PASS colorComponents[2] is 0
PASS colorComponents[3] is 0
PASS colorComponents[1] is 255
PASS colorComponents[2] is 0
PASS colorComponents[3] is 0
PASS colorComponents[1] is 0
PASS colorComponents[2] is 255
PASS colorComponents[3] is 255
PASS colorComponents[1] is 0
PASS colorComponents[2] is 255
PASS colorComponents[3] is 255
PASS colorComponents[1] is 128
PASS colorComponents[2] is 255
PASS colorComponents[3] is 255
PASS successfullyParsed is true

TEST COMPLETE
